Since we have no reason to reInit the javascript (no more SPF) lets simplify each module and remove the complexity of registering each one as a global
