+++ resources/js/main.js
@@ -2,7 +2,6 @@
 import 'intersection-observer';
 
 // Modules
-import './modules/app';
 import './modules/table';
 import './modules/sticky';
 import './modules/formfilter';
@@ -10,6 +9,7 @@ import './modules/media';
 import './modules/share';
 import './modules/lazy-image-loading';
 import './modules/anchor';
-import './modules/accordions';
-import './modules/offcanvas';
+import './modules/accordion';
+import './modules/slideout';
+import './modules/slideout-main-menu';
 import './modules/carousel';
