+++ resources/js/modules/anchor.js
@@ -1,42 +1,23 @@
 (function() {
     "use strict";
 
-    /**
-     * Anchor module.
-     */
-    class Anchor {
-        constructor() {
-            this._init();
-        }
+    var scrollToHash = function () {
+        var hash = window.location.hash.substring(1);
 
-        /**
-         * Initialize
-         */
-        _init() {
-            // Wait until the page finishes loading so the animate scrollTop doesn't bounce to the top of the pages
-            window.onload = function () {
-                window.WayneState.modules.anchor.scrollToHash();
-            }
+        if(hash !== '' && document.getElementById(hash)) {
+            let y = document.getElementById(hash).offsetTop - document.querySelector('.menu-top-container').offsetHeight;
 
-
-            // Anytime the hash changes after the page is loaded scroll the user to that hash
-            if ("onhashchange" in window) {
-                window.onhashchange = this.scrollToHash;
-            }
-        }
-
-        scrollToHash() {
-            var hash = window.location.hash.substring(1);
-
-            if(hash !== '' && document.getElementById(hash)) {
-                let y = document.getElementById(hash).offsetTop - document.querySelector('.menu-top-container').offsetHeight;
-
-                window.scroll(0, y);
-            }
+            window.scroll(0, y);
         }
     }
 
+    // Wait until the page finishes loading so the animate scrollTop doesn't bounce to the top of the pages
+    window.onload = function () {
+        scrollToHash();
+    }
 
-    // Register this module
-    window.WayneState.register('anchor', new Anchor());
+    // Anytime the hash changes after the page is loaded scroll the user to that hash
+    if ("onhashchange" in window) {
+        window.onhashchange = scrollToHash;
+    }
 })();
