+++ resources/js/modules/media.js
@@ -1,28 +1,7 @@
-import mediabox from 'mediabox';
+import Mediabox from 'mediabox';
 
-(function() {
+(function(Mediabox) {
     "use strict";
 
-    /**
-     * Media module.
-     */
-    class Media {
-
-        constructor() {
-            this._init();
-        }
-
-        /**
-         * Initialize
-         */
-        _init() {
-            mediabox('a[href*="youtu.be"], a[href*="youtube.com/watch"]');
-        }
-    }
-
-    // Initialize
-    var media = new Media();
-
-    // Register this module
-    window.WayneState.register('media', media);
-})();
+    Mediabox('a[href*="youtu.be"], a[href*="youtube.com/watch"]');
+})(Mediabox);
